Reader calculate Matrix

+++ src/Models/DataSet.kt
@@ -2,5 +2,5 @@ package Models
 
 data class DataSet (
     val nodes: List<Node>,
-    val matrix: Array<IntArray>
+    val matrix: Array<DoubleArray>
 )
